move tests for `set_meta` to own file (preparation for more tests)
